+++ ricaun.Revit.Installation.Tests/RevitUtils_Tests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using System;
+using System.Reflection;
 
 namespace ricaun.Revit.Installation.Tests
 {
@@ -19,6 +20,14 @@ namespace ricaun.Revit.Installation.Tests
             Console.WriteLine(RevitUtils.GetAllUsersApplicationPluginsFolder());
             Console.WriteLine(RevitUtils.GetAllUsersAddInFolder());
             Console.WriteLine(RevitUtils.GetAllUsersAddInFolder(2021));
+        }
+
+        [Test]
+        public void RevitUtils_Test_GetReferences()
+        {
+            var assemblyFile = Assembly.GetExecutingAssembly().Location;
+            Assert.IsFalse(RevitUtils.TryGetRevitVersion(assemblyFile, out int revitVersion));
+            Assert.That(revitVersion, Is.EqualTo(0));
         }
     }
 }